requirements.txt
click
colorama
pillow
psd-tools
